apps/file-manager/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
zhipuai>=2.1.0
pydantic>=2.7.0
pdfminer.six>=20221105
pypdf>=4.2.0
python-docx>=1.1.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
